Update base env + data gather correction (#4)
* scripts/data_gather_teleop: update rgb to uint8

* envs/base: update neutral orientation

+++ scripts/data_gather_teleop.py
@@ -101,7 +101,7 @@ class Dataset:
             actions.append(keystep["action"])
 
         outs = {
-                "rgb": torch.stack(rgbs).float().numpy(),
+                "rgb": torch.stack(rgbs).numpy().astype(np.uint8),
                 "pc": torch.stack(pcs).float().numpy(),
                 "gripper_uv": gripper_uv,
                 "action": np.stack(actions).astype(np.float32),
